move supg function/struct
