+++ include/mm_fill_util.h
@@ -27,6 +27,8 @@
 #define EXTERN extern
 #endif
 
+#include "mm_as_structs.h"
+
 EXTERN int beer_belly
 PROTO((void));
 
@@ -222,6 +224,11 @@ extern void determine_ProjectionVar(PROBLEM_DESCRIPTION_STRUCT *);
 extern void set_solid_inertia(void);
 
 extern int fill_variable_vector(int inode, int ivec_varType[], int ivec_matID[]);
+
+EXTERN void get_supg_tau(struct SUPG_terms *supg_terms,
+                         int dim,
+                         dbl diffusivity,
+                         PG_DATA *pg_data);
 
 
 #endif /* _MM_FILL_UTIL_H */
