+++ src/mm_fill_species.c
@@ -114,121 +114,6 @@
 /******************************************************************************/
 /******************************************************************************/
 
-struct SUPG_terms {
-  dbl supg_tau;
-  dbl d_supg_tau_dv[DIM][MDE];
-  dbl d_supg_tau_dX[DIM][MDE];
-};
-
-void get_supg_tau(struct SUPG_terms *supg_terms,
-                  int dim,
-                  dbl diffusivity,
-                  PG_DATA *pg_data)
-{
-  double vnorm = 0;
-
-  for (int i = 0; i < VIM; i++) {
-    vnorm += fv->v[i]*fv->v[i];
-  }
-  vnorm = sqrt(vnorm);
-
-  double hk = 0;
-  for (int i = 0; i < dim; i++) {
-    hk += sqrt(pg_data->hsquared[i]);
-  }
-
-  hk /= (double) dim;
-
-  double D = diffusivity;
-
-  double hk_dX[DIM][MDE];
-  for (int a = 0; a < dim; a++)
-    {
-      for (int j = 0; j < ei->dof[MESH_DISPLACEMENT1+a]; j++)
-        {
-          double tmp = 0;
-          for (int b = 0; b < dim; b++)
-            {
-              tmp += (2*pg_data->hhv[b][a] * pg_data->dhv_dxnode[b][j])/(2*sqrt(pg_data->hsquared[b]));
-            }
-          hk_dX[a][j] = tmp/dim;
-        }
-    }
-
-  double Pek = 0.5 * vnorm * hk / D;
-
-  double eta = Pek;
-  double eta_dX[DIM][MDE];
-  double eta_dV[DIM][MDE];
-  if (Pek > 1) {
-    eta = 1;
-    for (int i = 0; i < DIM; i++)
-    {
-      for (int j = 0; j < MDE; j++)
-      {
-        eta_dX[i][j] = 0;
-        eta_dV[i][j] = 0;
-      }
-    }
-  }
-  else
-  {
-    for (int i = 0; i < DIM; i++)
-    {
-      for (int j = 0; j < MDE; j++)
-      {
-        if (pd->e[VELOCITY1+i])
-        {
-          eta_dV[i][j] = 0.5 * 0.5 * hk * fv->v[i]*bf[VELOCITY1+i]->phi[j] / (vnorm*D);
-
-        }
-
-        if (pd->e[MESH_DISPLACEMENT1+i])
-        {
-          eta_dX[i][j] = 0.5 * vnorm * hk_dX[i][j] / D;
-
-        }
-      }
-    }
-  }
-
-  if (vnorm > 0) {
-    supg_terms->supg_tau = 0.5 * hk * eta / vnorm;
-
-    for (int a = 0; a < VIM; a++)
-      {
-        int var = VELOCITY1 + a;
-        for (int j = 0; j < ei->dof[var]; j++)
-          {
-            supg_terms->d_supg_tau_dv[a][j] = 0.5*hk*eta*fv->v[a]*bf[var]->phi[j] /
-                (- vnorm*vnorm*vnorm) + 0.5 * hk * eta_dV[a][j] / vnorm;
-          }
-
-        var = MESH_DISPLACEMENT1 + a;
-        for (int j = 0; j < ei->dof[var]; j++)
-          {
-            supg_terms->d_supg_tau_dX[a][j] = 0.5 * hk_dX[a][j] * eta / vnorm + 0.5 * hk * eta_dX[a][j] / vnorm;
-          }
-      }
-
-
-  } else {
-    supg_terms->supg_tau = 0;
-    for (int i = 0; i < DIM; i++)
-      {
-        for (int j = 0; j < MDE; j++)
-          {
-            supg_terms->d_supg_tau_dv[i][j] = 0.0;
-          }
-        for (int j = 0; j < MDE; j++)
-          {
-            supg_terms->d_supg_tau_dX[i][j] = 0.0;
-          }
-      }
-  }
-
-}
-
 /* assemble_mass_transport() -- assemble terms (Residual &| Jacobian) for mass
  *                              conservation eqns
  *
